requirements.txt
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.2.6
packaging==26.0
pillow==12.1.0
pycryptodome==3.23.0
pyparsing==3.3.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.18.0
pyqtgraph==0.14.0
python-dateutil==2.9.0.post0
six==1.17.0
